+++ src/test/java/kr/hhplus/be/server/coupon/CouponIssueTest.java
@@ -1,8 +1,10 @@
-package kr.hhplus.be.server.coupon.domain;
+package kr.hhplus.be.server.coupon;
 
 
 import kr.hhplus.be.server.common.exception.ApiErrorCode;
 import kr.hhplus.be.server.common.exception.ApiException;
+import kr.hhplus.be.server.coupon.domain.Coupon;
+import kr.hhplus.be.server.coupon.domain.CouponIssue;
 import kr.hhplus.be.server.user.domain.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
